tentando criar pacotes para os players
